requirements.txt
astroid==2.0.4
atomicwrites==1.2.1
attrs==18.2.0
autopep8==1.4
cachetools==3.0.0
certifi==2018.10.15
chardet==3.0.4
click==6.7
confusable-homoglyphs==3.2.0
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.1
coveralls==1.5.1
defusedxml==0.5.0
dj-database-url==0.5.0
Django==2.1.2
django-allauth==0.38.0
django-braces==1.13.0
django-cors-headers==2.4.0
django-cors-middleware==1.3.1
django-extensions==2.1.3
django-filter==2.0.0
django-heroku==0.3.1
django-nose==1.4.6
django-oauth-toolkit==1.2.0
django-registration==3.0
django-rest-auth==0.9.3
django-rest-framework-social-oauth2==1.1.0
django-rest-swagger==2.2.0
django-taggit==0.23.0
django-taggit-serializer==0.1.7
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
docopt==0.6.2
drf-yasg==1.11.0
facebook-sdk==3.0.0
flake8==3.6.0
flex==6.13.2
google-api-python-client==1.7.4
google-auth==1.5.1
google-auth-httplib2==0.0.3
gunicorn==19.9.0
httplib2==0.11.3
idna==2.7
inflection==0.3.1
isort==4.3.4
itypes==1.1.0
Jinja2==2.10
jsonpointer==1.14
jsonschema==2.6.0
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
mock==2.0.0
mock-django==0.6.10
more-itertools==4.3.0
nose==1.3.7
oauthlib==2.1.0
openapi-codec==1.3.2
pathspec==0.5.9
pbr==5.1.1
pluggy==0.8.0
psycopg2==2.7.5
psycopg2-binary==2.7.5
py==1.7.0
pyaml==17.12.1
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycodestyle==2.4.0
pyflakes==2.0.0
PyJWT==1.6.4
pylint==2.1.1
pytest==3.9.2
python-decouple==3.1
python-http-client==3.1.0
python3-openid==3.1.0
pytz==2018.5
PyYAML==3.13
requests==2.20.0
requests-oauthlib==1.0.0
rest-social-auth==1.4.0
rfc3987==1.3.8
rsa==4.0
ruamel.yaml==0.15.75
sendgrid==2.2.1
simplejson==3.16.0
six==1.11.0
smtpapi==0.3.1
social-auth-app-django==2.1.0
social-auth-core==1.7.0
strict-rfc3339==0.7
swagger-spec-validator==2.4.1
uritemplate==3.0.0
urllib3==1.24
validate-email==1.3
whitenoise==4.1
wrapt==1.10.11
yamllint==1.12.1
